wallet: extract out utility functions from wallet/wallet
To wallet/walletutil.h

Couple of functions had to be made non-static for them to be used by
wallet/wallet.

This patch should be reviewed with --color-moved=dimmed-zebra option.
