+++ src/wallet/wallet.cpp
@@ -94,54 +94,6 @@ using util::ToString;
 
 namespace wallet {
 
-bool AddWalletSetting(interfaces::Chain& chain, const std::string& wallet_name)
-{
-    const auto update_function = [&wallet_name](common::SettingsValue& setting_value) {
-        if (!setting_value.isArray()) setting_value.setArray();
-        for (const auto& value : setting_value.getValues()) {
-            if (value.isStr() && value.get_str() == wallet_name) return interfaces::SettingsAction::SKIP_WRITE;
-        }
-        setting_value.push_back(wallet_name);
-        return interfaces::SettingsAction::WRITE;
-    };
-    return chain.updateRwSetting("wallet", update_function);
-}
-
-bool RemoveWalletSetting(interfaces::Chain& chain, const std::string& wallet_name)
-{
-    const auto update_function = [&wallet_name](common::SettingsValue& setting_value) {
-        if (!setting_value.isArray()) {
-            if (wallet_name.empty() && setting_value.isNull()) {
-                // Empty setting suppresses backwards-compatible default wallet autoload.
-                setting_value.setArray();
-                return interfaces::SettingsAction::WRITE;
-            }
-            return interfaces::SettingsAction::SKIP_WRITE;
-        }
-        common::SettingsValue new_value(common::SettingsValue::VARR);
-        for (const auto& value : setting_value.getValues()) {
-            if (!value.isStr() || value.get_str() != wallet_name) new_value.push_back(value);
-        }
-        if (new_value.size() == setting_value.size()) return interfaces::SettingsAction::SKIP_WRITE;
-        setting_value = std::move(new_value);
-        return interfaces::SettingsAction::WRITE;
-    };
-    return chain.updateRwSetting("wallet", update_function);
-}
-
-static void UpdateWalletSetting(interfaces::Chain& chain,
-                                const std::string& wallet_name,
-                                std::optional<bool> load_on_startup,
-                                std::vector<bilingual_str>& warnings)
-{
-    if (!load_on_startup) return;
-    if (load_on_startup.value() && !AddWalletSetting(chain, wallet_name)) {
-        warnings.emplace_back(Untranslated("Wallet load on startup setting could not be updated, so wallet may not be loaded next node startup."));
-    } else if (!load_on_startup.value() && !RemoveWalletSetting(chain, wallet_name)) {
-        warnings.emplace_back(Untranslated("Wallet load on startup setting could not be updated, so wallet may still be loaded next node startup."));
-    }
-}
-
 /**
  * Refresh mempool status so the wallet is in an internally consistent state and
  * immediately knows the transaction's status: Whether it can be considered
@@ -2921,45 +2873,6 @@ bool CWallet::EraseAddressReceiveRequest(WalletBatch& batch, const CTxDestinatio
     if (!batch.EraseAddressReceiveRequest(dest, id)) return false;
     m_address_book[dest].receive_requests.erase(id);
     return true;
-}
-
-util::Result<fs::path> GetWalletPath(const std::string& name)
-{
-    const fs::path name_path = fs::PathFromString(name);
-
-    // 'name' must be a normalized path, i.e. no . or .. except at the root
-    if (name_path != name_path.lexically_normal()) {
-        return util::Error{Untranslated("Wallet name given as a path must be normalized")};
-    }
-
-    // 'name' cannot begin with ./ or ../
-    if (!name_path.empty() && (*name_path.begin() == fs::PathFromString(".") || *name_path.begin() == fs::PathFromString(".."))) {
-        return util::Error{Untranslated("Wallet name given as a relative path cannot begin with ./ or ../, for wallets not in the walletdir, please use an absolute path.")};
-    }
-
-    // Disallow path at root
-    if (name_path.has_root_path() && name_path.root_path() == name_path) {
-        return util::Error{Untranslated("Wallet name cannot be the root path")};
-    }
-
-    // Do some checking on wallet path. It should be either a:
-    //
-    // 1. Path where a directory can be created.
-    // 2. Path to an existing directory.
-    // 3. Path to a symlink to a directory.
-    // 4. For backwards compatibility, the name of a data file in -walletdir.
-    const fs::path wallet_path = fsbridge::AbsPathJoin(GetWalletDir(), name_path);
-    fs::file_type path_type = fs::symlink_status(wallet_path).type();
-    if (!(path_type == fs::file_type::not_found || path_type == fs::file_type::directory ||
-          (path_type == fs::file_type::symlink && fs::is_directory(wallet_path)) ||
-          (path_type == fs::file_type::regular && name_path.filename() == name_path))) {
-        return util::Error{Untranslated(strprintf(
-              "Invalid -wallet path '%s'. -wallet path should point to a directory where wallet.dat and "
-              "database/log.?????????? files can be stored, a location where such a directory could be created, "
-              "or (for backwards compatibility) the name of an existing data file in -walletdir (%s)",
-              name, fs::quoted(fs::PathToString(GetWalletDir()))))};
-    }
-    return wallet_path;
 }
 
 std::unique_ptr<WalletDatabase> MakeWalletDatabase(const std::string& name, const DatabaseOptions& options, DatabaseStatus& status, bilingual_str& error_string)
