+++ src/wallet/walletutil.h
@@ -5,6 +5,7 @@
 #ifndef BITCOIN_WALLET_WALLETUTIL_H
 #define BITCOIN_WALLET_WALLETUTIL_H
 
+#include <interfaces/chain.h>
 #include <script/descriptor.h>
 #include <util/fs.h>
 
@@ -58,6 +59,19 @@ enum WalletFlags : uint64_t {
 
 //! Get the path of the wallet directory.
 fs::path GetWalletDir();
+
+util::Result<fs::path> GetWalletPath(const std::string& name);
+
+//! Add wallet name to persistent configuration so it will be loaded on startup.
+bool AddWalletSetting(interfaces::Chain& chain, const std::string& wallet_name);
+
+//! Remove wallet name from persistent configuration so it will not be loaded on startup.
+bool RemoveWalletSetting(interfaces::Chain& chain, const std::string& wallet_name);
+
+void UpdateWalletSetting(interfaces::Chain& chain,
+                                const std::string& wallet_name,
+                                std::optional<bool> load_on_startup,
+                                std::vector<bilingual_str>& warnings);
 
 /** Descriptor with some wallet metadata */
 class WalletDescriptor
